+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-## Upcoming
+## v0.5.1 2016-12-25
 
 * Added `K.getValue()` function.
 * Added `K.promiseProperty()` function.